added check if anime exists

+++ main.py
@@ -13,6 +13,7 @@ while True:
     print("5. print all animes in table")
 
     choice = input("Pick 1-5 to do something: ").strip()
+    print("\n")
   
 
     match choice:
@@ -38,9 +39,17 @@ while True:
 
                 print("Anime shown successfully.")
                 print(this_anime)
+                print("\n")
+
+#-----------------------------------------------------------------------
 
             case "2":  #update an anime 
                anime_name = input("Enter the anime name: ").strip()
+               
+               if anime.anime_real(anime_name) == False: #check if the anime exists
+                    print("anime not found")
+                    continue
+               
                
                try:    
                     watched_episodes = int(input("Enter the number of episodes you have watched: "))
@@ -55,30 +64,56 @@ while True:
                updated_anime = anime.update_anime(anime_name,watched_episodes)
                print("anime updated sucessfully")
                print(updated_anime)
+               print("\n")
+               
+#-----------------------------------------------------------------------
 
             case "3": #delete an anime
                 anime_name = input("Enter the anime name: ").strip()
+
+                if anime.anime_real(anime_name) == False: #check if the anime exists
+                    print("anime not found")
+                    continue
+
+
                 deleteanime = anime.delete_anime(anime_name)
                 print("anime deleted")
                 print(deleteanime)
+                print("\n")
 
+
+#-----------------------------------------------------------------------
             case "4": #whats the progress on all anime?
                 anime_name = input("Enter the anime name: ").strip()
+
+                if anime.anime_real(anime_name) == False:  #check if the anime exists
+                    print("anime not found")
+                    continue
+
+
                 progressanime = anime.anime_progress(anime_name)
                 completedanime = anime.anime_completed(anime_name)
                 print (progressanime)
                 print (completedanime)
+                print("\n")
+
+#-----------------------------------------------------------------------
 
             case "5":  #print anime name list
                 anime.anime_titles()
                 print("\n")
 
+
+#-----------------------------------------------------------------------
+
             case default:
                 print("use numbers punk ass bitch")
+                print("\n")
 
 
     continue_choice = input("Do you want to continue adding anime? (Y/N): ").strip().lower()
     if continue_choice != ("y"):
         print("no more anime added")
+        print("\n")
         break
